+++ NEWS.md
@@ -7,6 +7,9 @@
 
 ## Features / Fixes
 
+* Fixed a rare issue where the deprecated `.progress` bar may cause an integer
+  overflow with extremely long inputs (#288).
+
 * Detangled furrr's documentation from purrr's to avoid some documentation
   inheritance issues (#286).